fix(input): clamp scroll deltas at the wire boundary (RemEx-hnin)
A mouseScroll delta went from the wire to the input backends unclamped, and
int.MinValue reached Math.Abs, which has no representable result and throws
rather than saturating. The throw is not the interesting part; what it cost
is.

RemoteDesktopHandler.DispatchInput catches Win32Exception,
InvalidOperationException and ArgumentException. OverflowException derives
from ArithmeticException and is none of them, so it escaped into
ProcessInputQueue -- whose own catch list is narrower still -- and ended its
GetConsumingEnumerable loop. That thread is started once in the constructor
and nothing restarts it, so every later mouse and keyboard event for the
session was queued and never drained while the video kept streaming. Silent
at every layer: the faulted task is observed only in Dispose, by a catch of
AggregateException whose whole body is the comment "expected".

Fixed at the boundary rather than in the branch that threw, because that is
the only place covering all four backends at once. Both dispatchers now go
through a new CoordinateValidation.ClampScrollDelta, joining the absolute
coordinates already clamped by ClampToActiveBounds and the pointer deltas
already clamped by ClampDelta. No wire change and no protocolVersion
implication -- the format is byte-identical and a host clamping untrusted
input is not a contract change.

The bound is ten detents, which is what both LinuxInputSimulationService
scroll branches already saturate at, so it is invisible there. It exists for
the two that bounded nothing: Windows passes the value straight to
MOUSEEVENTF_WHEEL, and the backend router's xdotool fallback spawned one
process per detent with no ceiling at all -- hence a magnitude bound rather
than an int.MinValue special case. Both Math.Abs sites now widen to long
first, as defence in depth for a caller reaching a backend without a handler.

Review corrected four over-broad claims in the first draft's comments, three
of them inherited from the bead's own wording: PingPongHandler dispatches
inline behind a catch (Exception), so only the remote-desktop dispatcher
could lose a session; the router's fallback was not among the paths that
"already saturate"; and LinuxInputSimulationService's xdotool branch negates
rather than taking a magnitude, and unary minus on int.MinValue is unchecked,
so it never threw -- it quietly emitted one click instead of ten.

Proven by injection: restoring the original code reproduces the real
OverflowException and fails 4 tests; widening the bound fails the test that
pins it. Not pinned, and said so rather than hidden: PingPongHandler's clamp
has no test (14-arg constructor, five concrete types), and the "session dies
permanently" half is verified by reading. The containment gap -- one bad
event should cost that event, not the session -- is RemEx-q4wm.

1267 tests green on Windows, 0 warnings; Android release build clean, which
is also the NativeAOT link check for the Remex.Core addition.

+++ remex.agent/Services/Input/LinuxInputSimulationService.cs
@@ -394,7 +394,7 @@ public class LinuxInputSimulationService : IInputSimulationService
     /// single message cannot fling the page.
     /// </remarks>
     internal static int WheelDetents(int delta) =>
-        delta == 0 ? 0 : Math.Sign(delta) * Math.Clamp(Math.Abs(delta) / 120, 1, 10);
+        delta == 0 ? 0 : Math.Sign(delta) * (int)Math.Clamp(Math.Abs((long)delta) / 120, 1, 10);
 
     /// <summary>
     /// Synchronously ensures the portal input session is active. The first caller blocks
